requirements.txt
fastapi
uvicorn
chromadb
google-genai
yt-dlp
pydantic
python-multipart